Fixed Error When Text File Doesn't Exist
I fixed an issue where if the file didn't exist, it would create an error. This was because the line in the timer method to WriteAllText to the file was conflicting with the line to create the file if the file didn't exist. By deleting the segment that created the file if it didn't exist, I resolved the issue. I really didn't need it, since WriteAllText creates the file if it doesn't exist anyways.